ConsistencyLoss/motionbertH36m/requirements.txt
tensorboardX
tqdm
easydict
prettytable
chumpy
opencv-python
imageio-ffmpeg
# matplotlib==3.1.1
roma
ipdb
pytorch-metric-learning                         # For one-shot action recognition
smplx[all]                                      # For mesh recovery
